GCP/requirements.txt
apispec==6.8.4
appengine-python-standard==1.1.10
atomicwrites==1.4.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.2
frozendict==2.4.6
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
legacy-cgi==2.6.3
MarkupSafe==3.0.3
marshmallow==4.0.1
mock==5.2.0
more-itertools==10.8.0
opentelemetry-api==1.37.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==1.26.20
wcwidth==0.2.14
webargs==8.7.0
Werkzeug==3.1.3
zipp==3.23.0
